webscraping/Cycle II - Scrapy Refactor and ETL/requirements.txt
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cssselect==1.1.0
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
elastic-transport==8.1.2
elasticsearch==8.3.3
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.9.1
fake-useragent==0.1.11
fastjsonschema==2.16.1
fonttools==4.34.4
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.8.0
jupyter-client==7.3.4
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions @ https://github.com/ipython-contrib/jupyter_contrib_nbextensions/tarball/master
jupyter-core==4.11.1
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.5.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lxml==4.9.1
manufacturing==1.3.3
marisa-trie==0.7.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numpy==1.23.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
pyee==8.2.2
Pygments==2.12.0
PyMySQL==1.0.2
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
regex==2022.7.25
requests==2.28.1
requests-html==0.10.0
scipy==1.9.0
selenium==4.3.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
stack-data==0.3.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
w3lib==1.22.0
wcwidth==0.2.5
webdriver-manager==3.8.3
webencodings==0.5.1
websockets==10.3
wsproto==1.1.0
xlrd==2.0.1
zipp==3.8.1
